pyproject.toml
[tool.poetry]
name = "scotty"
version = "2.4.9"
description = ""
authors = ["Dor Amram <damram@infinidat.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.6"
alembic = "^1.1"
amqp = "^2.5"
anyjson = "^0.3.3"
argh = "^0.26.2"
arrow = "^0.15.2"
async-timeout = "^3.0"
atomicwrites = "^1.3"
attrs = "^19.1"
backports-abc = "^0.5.0"
billiard = "^3.6"
blinker = "^1.4"
celery = "^4.3"
certifi = "^2019.9"
chardet = "^3.0"
click = "^7.0"
colorama = "^0.4.1"
confetti = "^2.5"
contextlib2 = "^0.5.5"
decorator = "^4.4"
defusedxml = "^0.6.0"
dessert = "^1.3"
ecdsa = "^0.13.2"
Flask = "^1.1"
Flask-Login = "^0.4.1"
Flask-Loopback = "^1.4"
Flask-Mail = "^0.9.1"
Flask-Migrate = "^2.5"
Flask-Principal = "^0.4.0"
Flask-Script = "^2.0"
Flask-Security = "^3.0"
Flask-SQLAlchemy = "^2.4"
Flask-WTF = "^0.14.2"
flux = "^1.3"
google-api-python-client = "^1.7"
gossip = "^2.4"
gunicorn = "^19.9"
httplib2 = "^0.13.1"
idna = "^2.8"
importlib-metadata = "^0.23.0"
ipython = "^7.8"
ipython-genutils = "^0.2.0"
itsdangerous = "^1.1"
Jinja2 = "^2.10"
jira = "^2.0"
jsonschema = "^3.0"
kaptan = "^0.5.12"
kombu = "^4.6"
libtmux = "^0.8.2"
livereload = "^2.6"
Logbook = "^1.5"
Mako = "^1.1"
MarkupSafe = "^1.1"
more-itertools = "^7.2"
multidict = "^4.5"
munch = "^2.3"
oauth2client = "^4.1"
oauthlib = "^3.1"
orderedset = "^2.0"
packaging = "^19.2"
paramiko = "^2.6"
passlib = "^1.7"
pathtools = "^0.1.2"
pbr = "^5.4"
pexpect = "^4.7"
pickleshare = "^0.7.5"
pluggy = "^0.13.0"
prompt-toolkit = "^2.0"
psutil = "^5.6"
psycopg2 = "^2.8"
psycopg2-binary = "^2.8"
ptyprocess = "^0.6.0"
py = "^1.8"
pyasn1 = "^0.4.7"
pyasn1-modules = "^0.2.6"
pycparser = "^2.19"
pycrypto = "^2.6"
Pygments = "^2.4"
pyparsing = "^2.4"
python-dateutil = "^2.8"
python-editor = "^1.0"
pytz = "^2019.2"
PyYAML = "^5.1"
raven = "^6.10"
redis = "^3.3"
requests = "^2.22"
requests-oauthlib = "^1.2"
requests-toolbelt = "^0.9.1"
rsa = "^4.0"
sentinels = "^1.0"
simplegeneric = "^0.8.1"
simplejson = "^3.16"
six = "^1.12"
SQLAlchemy = "^1.3.8"
tlslite = "^0.4.9"
tmuxp = "^1.5"
tornado = "^6.0"
traitlets = "^4.3"
uritemplate = "^3.0"
urllib3 = "^1.25"
URLObject = "^2.4"
urwid = "^2.0"
watchdog = "^0.9.0"
wcwidth = "^0.1.7"
weber-utils = "^1.2"
Werkzeug = "^0.15.6"
WTForms = "^2.2"
yarl = "^1.3"
zipp = "^0.6.0"
emport = "^1.2"
slash = "1.0.2"
pip = "^19.2"

[tool.poetry.dev-dependencies]
pudb = "^2019.1"
mypy = "^0.730.0"
pact = "^1.12"
pytest = "^5.2"
scottypy = "^0.20.0"
pytest-flask-sqlalchemy = "^1.0.2"
isort = "^4.3.21"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
